% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conseguiR-package.R
\docType{package}
\name{conseguiR-package}
\alias{conseguiR}
\alias{conseguiR-package}
\title{conseguiR package metadata}
\description{
Minimal package-level roxygen block used to declare the runtime namespaces
that \code{conseguiR} depends on across its exported wrapper layer and sourced
internal scripts.
}
\author{
\strong{Maintainer}: Pranav Mishra \email{pranav.mishra@yale.edu}

Authors:
\itemize{
  \item Mark Robinson \email{mark.robinson@yale.edu}
}

}
